[Feat/#268] refactor error status name was executed when admin user remove data
