+++ src/main/java/com/example/waggle/domain/board/application/siren/SirenCommandServiceImpl.java
@@ -104,7 +104,7 @@ public class SirenCommandServiceImpl implements SirenCommandService {
     @Override
     public void deleteSirenByAdmin(Long boardId, Member member) {
         if (!member.getRole().equals(Role.ADMIN)) {
-            throw new MemberHandler(ErrorStatus.MEMBER_REQUEST_IS_UNACCEPTABLE_BECAUSE_OF_AUTHORIZATION);
+            throw new MemberHandler(ErrorStatus.MEMBER_ACCESS_DENIED_BY_AUTHORIZATION);
         }
         commentRepository.deleteCommentsWithRelationsByBoard(boardId);
         boardRepository.deleteBoardsWithRelations(SIREN, boardId);
