+++ src/main/java/com/example/waggle/domain/board/application/story/StoryCommandServiceImpl.java
@@ -95,7 +95,7 @@ public class StoryCommandServiceImpl implements StoryCommandService {
     @Override
     public void deleteStoryByAdmin(Long boardId, Member member) {
         if (!member.getRole().equals(Role.ADMIN)) {
-            throw new MemberHandler(ErrorStatus.MEMBER_REQUEST_IS_UNACCEPTABLE_BECAUSE_OF_AUTHORIZATION);
+            throw new MemberHandler(ErrorStatus.MEMBER_ACCESS_DENIED_BY_AUTHORIZATION);
         }
         commentRepository.deleteCommentsWithRelationsByBoard(boardId);
         boardRepository.deleteBoardsWithRelations(STORY, boardId);
