+++ src/main/java/com/example/waggle/domain/conversation/application/comment/CommentCommandServiceImpl.java
@@ -95,7 +95,7 @@ public class CommentCommandServiceImpl implements CommentCommandService {
     @Override
     public void deleteCommentByAdmin(Long commentId, Member member) {
         if (!member.getRole().equals(Role.ADMIN)) {
-            throw new MemberHandler(ErrorStatus.MEMBER_REQUEST_IS_UNACCEPTABLE_BECAUSE_OF_AUTHORIZATION);
+            throw new MemberHandler(ErrorStatus.MEMBER_ACCESS_DENIED_BY_AUTHORIZATION);
         }
         commentRepository.deleteCommentsWithRelations(commentId);
     }
