+++ src/main/java/com/example/waggle/domain/pet/application/PetCommandServiceImpl.java
@@ -54,7 +54,7 @@ public class PetCommandServiceImpl implements PetCommandService {
     @Override
     public void deletePetByAdmin(Long petId, Member member) {
         if (!member.getRole().equals(Role.ADMIN)) {
-            throw new MemberHandler(ErrorStatus.MEMBER_REQUEST_IS_UNACCEPTABLE_BECAUSE_OF_AUTHORIZATION);
+            throw new MemberHandler(ErrorStatus.MEMBER_ACCESS_DENIED_BY_AUTHORIZATION);
         }
         petRepository.deleteById(petId);
     }
