+++ src/main/java/com/example/waggle/domain/schedule/application/ScheduleQueryService.java
@@ -1,4 +1,4 @@
-package com.example.waggle.domain.schedule.application.schedule;
+package com.example.waggle.domain.schedule.application;
 
 import com.example.waggle.domain.member.persistence.entity.Member;
 import com.example.waggle.domain.schedule.persistence.entity.Schedule;
