+++ src/main/java/com/example/waggle/domain/schedule/presentation/controller/ScheduleApiController.java
@@ -3,8 +3,8 @@ package com.example.waggle.domain.schedule.presentation.controller;
 import com.example.waggle.domain.member.persistence.entity.Member;
 import com.example.waggle.domain.member.presentation.converter.MemberConverter;
 import com.example.waggle.domain.member.presentation.dto.MemberResponse.MemberSummaryListDto;
-import com.example.waggle.domain.schedule.application.schedule.ScheduleCommandService;
-import com.example.waggle.domain.schedule.application.schedule.ScheduleQueryService;
+import com.example.waggle.domain.schedule.application.ScheduleCommandService;
+import com.example.waggle.domain.schedule.application.ScheduleQueryService;
 import com.example.waggle.domain.schedule.persistence.entity.Schedule;
 import com.example.waggle.domain.schedule.presentation.converter.ScheduleConverter;
 import com.example.waggle.domain.schedule.presentation.dto.schedule.ScheduleRequest;
