+++ src/main/java/com/example/waggle/domain/schedule/presentation/controller/TeamApiController.java
@@ -3,8 +3,8 @@ package com.example.waggle.domain.schedule.presentation.controller;
 import com.example.waggle.domain.member.persistence.entity.Member;
 import com.example.waggle.domain.member.presentation.converter.MemberConverter;
 import com.example.waggle.domain.member.presentation.dto.MemberResponse.MemberSummaryListDto;
-import com.example.waggle.domain.schedule.application.team.TeamCommandService;
-import com.example.waggle.domain.schedule.application.team.TeamQueryService;
+import com.example.waggle.domain.schedule.application.TeamCommandService;
+import com.example.waggle.domain.schedule.application.TeamQueryService;
 import com.example.waggle.domain.schedule.persistence.entity.Participation;
 import com.example.waggle.domain.schedule.persistence.entity.Team;
 import com.example.waggle.domain.schedule.presentation.converter.TeamConverter;
