+++ src/main/java/com/example/waggle/exception/payload/code/ErrorStatus.java
@@ -1,16 +1,10 @@
 package com.example.waggle.exception.payload.code;
 
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
-import static org.springframework.http.HttpStatus.CONFLICT;
-import static org.springframework.http.HttpStatus.FORBIDDEN;
-import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
-import static org.springframework.http.HttpStatus.NOT_ACCEPTABLE;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-import static org.springframework.http.HttpStatus.UNAUTHORIZED;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
+
+import static org.springframework.http.HttpStatus.*;
 
 @Getter
 @AllArgsConstructor
@@ -52,7 +46,7 @@ public enum ErrorStatus implements BaseCode {
     MEMBER_DUPLICATE_NICKNAME(CONFLICT, 4106, "이미 사용중인 닉네임입니다."),
     MEMBER_NAME_TYPE_IS_INVALID(NOT_ACCEPTABLE, 4107, "랜덤 입력할 필드의 타입을 잘못 입력하였습니다"),
     MEMBER_EMAIL_VERIFICATION_FAILED(BAD_REQUEST, 4108, "인증번호가 일치하지 않습니다. 이메일 인증에 실패했습니다."),
-    MEMBER_REQUEST_IS_UNACCEPTABLE_BECAUSE_OF_AUTHORIZATION(BAD_REQUEST, 4109, "인가가 잘못된 요청입니다."),
+    MEMBER_ACCESS_DENIED_BY_AUTHORIZATION(BAD_REQUEST, 4109, "인가가 잘못된 요청입니다."),
 
 
     // 게시판 관련 오류 (4150 ~ 4199)
