+++ src/test/java/com/example/waggle/domain/notification/service/NotificationQueryServiceTest.java
@@ -6,7 +6,7 @@ import com.example.waggle.domain.member.persistence.entity.Member;
 import com.example.waggle.domain.notification.application.NotificationCommandService;
 import com.example.waggle.domain.notification.application.NotificationQueryService;
 import com.example.waggle.domain.notification.persistence.entity.Notification;
-import com.example.waggle.domain.schedule.application.team.TeamCommandService;
+import com.example.waggle.domain.schedule.application.TeamCommandService;
 import com.example.waggle.global.service.test.DatabaseCleanUp;
 import lombok.extern.slf4j.Slf4j;
 import org.assertj.core.groups.Tuple;
